# whareflow-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "WareFlow api",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8089",
    "basePath": "/api/v1/employer",
    "paths": {
...
        "/warehouse/{warehouse_id}/product/{product_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список всех товаров со склада",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Получение всех товаров со склада",
                "parameters": [
                    {
                        "type": "string",
                        "description": "warehouse id",
                        "name": "warehouse_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "[]delivery.ProductModelResponse",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error: invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обнвляет данные о продукте",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Обновление продукта",
                "parameters": [
                    {
                        "type": "string",
                        "description": "warehouse id",
                        "name": "warehouse_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "warehouse id",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для создания склада",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delivery.ProductModelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: product success updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error: invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/warehouse/{warehouse_id}/zone/{zone_id}/product": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список всех товаров с зоны склада",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Получение всех товаров с зоны склада",
                "parameters": [
                    {
                        "type": "string",
                        "description": "warehouse id",
                        "name": "warehouse_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "zone id",
                        "name": "zone_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "[]delivery.ProductModelResponse",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error: invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новый продукт",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Создание продукта",
                "parameters": [
                    {
                        "type": "string",
                        "description": "warehouse id",
                        "name": "warehouse_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "zone id",
                        "name": "zone_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для создания склада",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delivery.ProductModelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: product success created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error: invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/warehouse/{warehouse_id}/zone/{zone_id}/product/{product_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает товар",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Получение товара",
                "parameters": [
                    {
                        "type": "string",
                        "description": "warehouse id",
                        "name": "warehouse_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "zone id",
                        "name": "zone_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "product id",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delivery.ProductModelResponse",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error: invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "delivery.ProductModelRequest": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}